This may break the build but it's the only way I could build for distribution

+++ Classes/Data Sources/PartisanIndexStats.m
@@ -196,7 +196,7 @@
 	if (tempNum)
 		maxWnomSession = tempNum.integerValue;
 	
-	NSMutableString *predicateString = [NSMutableString stringWithFormat:@"self.legislator != nil AND self.legislator.legtype == %d AND self.session == %d", chamber, maxWnomSession];
+	NSMutableString *predicateString = [NSMutableString stringWithFormat:@"self.legislator != nil AND self.legislator.legtype == %d AND self.session == %d", chamber, (UInt16)maxWnomSession];
 	
 	if (party > BOTH_PARTIES)
 		[predicateString appendFormat:@" AND self.legislator.party_id == %d", party];
